Логер пишет до лимитера. Добавить переменные в env
